[Oxfordshire] Added report video to Oxfordshire Home and report page
On the homepage the video is not using fancy box, because the space/size
is big enough to actually play the video without the need of an
overlay.

For the video when reporting a problem, the file containing the video has
been created, however It still need to be added to the Oxfordshire
template with a conditional, that only includes the video when we are
visiting the Oxfordshire cobrand. It should be included above the
"Category list".